+++ core/src/python_wrapper/py_maiacore.cpp
@@ -3,6 +3,7 @@
 
 namespace py = pybind11;
 
+void ScoreCollectionClass(py::module &);
 void ScoreClass(py::module &);
 void PartClass(py::module &);
 void ClefClass(py::module &);
@@ -28,6 +29,7 @@ PYBIND11_MODULE(maiacore, m) {
     MeasureClass(m);
     IntervalClass(m);
     HelperClass(m);
+    ScoreCollectionClass(m);
     
 }
 #endif